Created some fundamental methods
Created a basic structure for our implementation for the game.
We will have to figure out if we want to implement the board with a list (8x8) containing the squares or if we want to work with enums (I think that would not be good).
Just have a look at the classes.